Added static class-level variables: - DATA_LINE_SEPARATOR for line-by-line separation of input data into operation type and amount; - DATA_BUILDER_SEPARATOR to separate the result of a file read and written by StringBuilder; - OPERATION_POSITION to read the operation type in a split line; - AMOUNT_POSITION for amount respectively.
In the getStatistic() method, a new File object is initialized, which is then passed to the readFromFile() method.

Added a readFromFile() method that reads a file using the BufferedReader and writes the result to a StringBuilder, after which it is passed to the calculateData() method before being passed to the writeToFile() method for writing.

The calculateData() method has been added, in which the supply and buy variables are declared for ease of reading. After that, a String[] array is declared, which is initialized by builder.toString(), split by DATA_BUILDER_SEPARATOR (" ", since the StringBuilder formed in the previous method contains spaces instead of a line separator).
Then we cycle through the new array. Inside the cycle, we create another auxiliary array, into which we pass the element of the first array, split by DATA_LINE_SEPARATOR (",").
Using if, we calculate the supply and buy values, knowing their exact indices.
After the cycle, we reset the StringBuilder, after which we add the calculated data to it, according to the template:
```
supply,__amount__
buy,__amount__
result,__amount__
```
We return StringBuilder.

The writeToFile() method has been added, which, after counting, creates a new File object, initializes it with the name toFileName, and writes a StringBuilder to it byte by byte.