The method opens the file fromFileName for reading using the BufferedReader class and reads it line by line in a loop. Each line contains two fields separated by a comma: operation and amount. The first field indicates the type of operation ("supply" - delivery, "buy" - purchase), the second field indicates the quantity of goods.
During the reading process, the method calculates the total quantity of the product that was supplied (totalSupply) and the total quantity of the product that was purchased (totalBuy).

When the file is finished reading, the method closes it and calculates the difference between totalSupply and totalBuy, writing this result to the result variable.